Remove subparsing to differentiate between draw options and sequence options
This is done to easily call drawer.py for a functional program without
having to trail the call with "seq <length of sequence>".

Now a default value (stored in reca_sequence.py) of 100 will be used if
value is omitted. To specify a sequence length, instead use either the
-sl or --seqlen flags. Considered converting all short-flags from
two-letter codes to one-letter codes, but failed to construct intuitive
combinations.